#17: implement ugit log

+++ ugit/base.py
@@ -1,5 +1,9 @@
-from ntpath import dirname
+import itertools
+import operator
+import operator
 import os
+
+from collections import namedtuple
 from . import data
 
 
@@ -56,11 +60,15 @@ def _empty_current_directory():
             if is_ignored(path) or not os.path.isfile(path):
                 continue
             os.remove(path)
-        try:
-            os.rmdir(path)
-        except (FileNotFoundError, OSError):
-            # Deletion might file in case of ignored files, but that's OK.
-            pass
+        for dirname in dirnames:
+            path = os.path.relpath(f"{root}/{dirname}")
+            if is_ignored(path):
+                continue
+            try:
+                os.rmdir(path)
+            except (FileNotFoundError, OSError):
+                # Deletion might file in case of ignored files, but that's OK.
+                pass
 
 
 def read_tree(tree_oid):
@@ -85,6 +93,27 @@ def commit(message):
     data.set_HEAD(oid)
 
     return oid
+
+
+Commit = namedtuple("Comit", ["tree", "parent", "message"])
+
+
+def get_commit(oid):
+    parent = None
+
+    commit = data.get_object(oid, "commit").decode()
+    lines = iter(commit.splitlines())
+    for line in itertools.takewhile(operator.truth, lines):
+        key, value = line.split(" ", 1)
+        if key == "tree":
+            tree = value
+        elif key == "parent":
+            parent = value
+        else:
+            assert False, f"Unknown field {key}"
+        
+    message = "\n".join(lines)
+    return Commit(tree=tree, parent=parent, message=message)
 
 
 def is_ignored(path):
